Aula 03/10/2023

+++ backend/src/controllers/topic.controller.ts
@@ -1,4 +1,4 @@
-import { Body, Controller, Delete, Get, HttpCode, Param, ParseIntPipe, Post } from "@nestjs/common";
+import { Body, Controller, Delete, Get, HttpCode, HttpException, HttpStatus, Param, ParseIntPipe, Post, Put } from "@nestjs/common";
 import { Topic } from "src/entities/topic.entity";
 import { TopicService } from "src/services/topic.service";
 
@@ -13,8 +13,14 @@ export class TopicController {
     }
 
     @Get(':id')
-    findById(@Param('id', ParseIntPipe) id: number): Promise<Topic> {
-        return this.service.findById(id)
+    async findById(@Param('id', ParseIntPipe) id: number): Promise<Topic> {
+        const found = await this.service.findById(id);
+
+        if(!found) {
+            throw new HttpException('Topic not found', HttpStatus.NOT_FOUND);
+        }
+
+        return found;
     }
 
     @Post()
@@ -24,8 +30,24 @@ export class TopicController {
 
     @Delete(':id')
     @HttpCode(204)
-    delete(@Param('id', ParseIntPipe) id: number): Promise<void>  {
+    async delete(@Param('id', ParseIntPipe) id: number): Promise<void>  {
+        const found = await this.service.findById(id);
+
+        if(!found) {
+            throw new HttpException('Topic not found', HttpStatus.NOT_FOUND);
+        }
+
         return this.service.delete(id);
     }
 
+    @Put(':id')
+    async update(@Param('id', ParseIntPipe) id: number, @Body() topic: Topic): Promise<Topic> {
+        const found = await this.service.findById(id);
+
+        if(!found) {
+            throw new HttpException('Topic not found', HttpStatus.NOT_FOUND);
+        }
+
+        return this.service.update(id, topic);
+    }
 }
