+++ backend/src/controllers/user.controller.ts
@@ -1,4 +1,4 @@
-import { Body, Controller, Delete, Get, HttpCode, Param, ParseIntPipe, Post } from "@nestjs/common";
+import { Body, Controller, Delete, Get, HttpCode, HttpException, HttpStatus, Param, ParseIntPipe, Post, Put } from "@nestjs/common";
 import { User } from "src/entities/user.entity";
 import { UserService } from "src/services/user.service";
 
@@ -13,8 +13,25 @@ export class UserController {
     }
 
     @Get(':id')
-    findById(@Param('id', ParseIntPipe) id: number): Promise<User> {
-        return this.service.findById(id)
+    async findById(@Param('id', ParseIntPipe) id: number): Promise<User> {
+        const found = await this.service.findById(id);
+
+        if(!found) {
+            throw new HttpException('User not found', HttpStatus.NOT_FOUND);
+        }
+
+        return found;
+    }
+
+    @Get(':username')
+    async findByUsername(@Param('username') username: string): Promise<User> {
+        const found = await this.service.findByUsername(username);
+
+        if(!found) {
+            throw new HttpException('User not found', HttpStatus.NOT_FOUND);
+        }
+
+        return found;
     }
 
     @Post()
@@ -24,8 +41,24 @@ export class UserController {
 
     @Delete(':id')
     @HttpCode(204)
-    delete(@Param('id', ParseIntPipe) id: number): Promise<void>  {
+    async delete(@Param('id', ParseIntPipe) id: number): Promise<void>  {
+        const found = await this.service.findById(id);
+
+        if(!found) {
+            throw new HttpException('User not found', HttpStatus.NOT_FOUND);
+        }
+
         return this.service.delete(id);
     }
 
+    @Put(':id')
+    async update(@Param('id', ParseIntPipe) id: number, @Body() user: User): Promise<User> {
+        const found = await this.service.findById(id);
+
+        if(!found) {
+            throw new HttpException('User not found', HttpStatus.NOT_FOUND);
+        }
+
+        return this.service.update(id, user);
+    }
 }
